requirements.txt
fastapi==0.68.1
uvicorn==0.13.4
motor==2.5.1
python-decouple
pydantic~=1.8.2
pymongo~=3.12.1